Part 4 Complete

+++ part4/blogbackend/controllers/blogs.js
@@ -1,6 +1,16 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+
+//--- HELPER FUNCTIONS ---
+const getToken = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+        return authorization.substring(7)
+      }
+      return null
+}
 
 //--- ENDPOINT: Info about Blogs ---
 blogsRouter.get('/info', (request, response) => {
@@ -17,12 +27,19 @@ blogsRouter.get('/', async (request, response) => {
 //--- ENDPOINT: Add a blog ---
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+
+    const token = getToken(request)    
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if(!token || !decodedToken.id){
+        return response.status(401).json({ error: 'Token Missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+
   
     if(!body.title && !body.url){
         response.status(400).end()
     }else{
-        const user = await User.findById(body.userId)
-
         const blog = new Blog({
             title: body.title,
             author: body.author,
@@ -74,6 +91,21 @@ blogsRouter.put('/:id', async (request, response) => {
 
 //--- ENDPOINT: Delete blog by ID ---
 blogsRouter.delete('/:id', async (request, response) => {
+    const body = request.body
+
+    const token = getToken(request)    
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if(!token || !decodedToken.id){
+        return response.status(401).json({ error: 'Token Missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() !== user.id.toString()){
+        return response.status(401).json({error: 'You are not authorized to delete this blog.'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
